requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.9
pydantic-settings>=2.5
httpx>=0.27
python-dotenv>=1.0
playwright>=1.49
beautifulsoup4>=4.12
lxml>=5.2
openai>=1.52
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0  # For testing scripts
